feat: 상품 등록/수정 시 thumbnail 등록/수정 기능 fix: Product entity id 타입 int->long 변경

+++ src/main/java/com/smplatform/product_service/domain/product/repository/ProductOptionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
-    List<ProductOption> findAllByProduct_Id(int productId);
+    List<ProductOption> findAllByProduct_Id(long productId);
 }
